[KYUUBI #4367] Rename delegation package to shim
